refactor: split settings orchestration into section hooks
