+++ src/features/settings/hooks/settingsSectionTypes.ts
@@ -1,0 +1,7 @@
+import type { WorkspaceInfo } from "@/types";
+
+export type GroupedWorkspaces = Array<{
+  id: string | null;
+  name: string;
+  workspaces: WorkspaceInfo[];
+}>;
