+++ src/features/settings/hooks/useSettingsCodexSection.ts
@@ -1,0 +1,355 @@
+import { useEffect, useState } from "react";
+import type { Dispatch, SetStateAction } from "react";
+import { open } from "@tauri-apps/plugin-dialog";
+import type {
+  AppSettings,
+  CodexDoctorResult,
+  CodexUpdateResult,
+  WorkspaceSettings,
+  WorkspaceInfo,
+} from "@/types";
+import { useGlobalAgentsMd } from "./useGlobalAgentsMd";
+import { useGlobalCodexConfigToml } from "./useGlobalCodexConfigToml";
+import { useSettingsDefaultModels } from "./useSettingsDefaultModels";
+import {
+  buildEditorContentMeta,
+  buildWorkspaceOverrideDrafts,
+} from "@settings/components/settingsViewHelpers";
+
+type UseSettingsCodexSectionArgs = {
+  appSettings: AppSettings;
+  projects: WorkspaceInfo[];
+  onUpdateAppSettings: (next: AppSettings) => Promise<void>;
+  onRunDoctor: (
+    codexBin: string | null,
+    codexArgs: string | null,
+  ) => Promise<CodexDoctorResult>;
+  onRunCodexUpdate?: (
+    codexBin: string | null,
+    codexArgs: string | null,
+  ) => Promise<CodexUpdateResult>;
+  onUpdateWorkspaceCodexBin: (id: string, codexBin: string | null) => Promise<void>;
+  onUpdateWorkspaceSettings: (
+    id: string,
+    settings: Partial<WorkspaceSettings>,
+  ) => Promise<void>;
+};
+
+export type SettingsCodexSectionProps = {
+  appSettings: AppSettings;
+  onUpdateAppSettings: (next: AppSettings) => Promise<void>;
+  defaultModels: ReturnType<typeof useSettingsDefaultModels>["models"];
+  defaultModelsLoading: boolean;
+  defaultModelsError: string | null;
+  defaultModelsConnectedWorkspaceCount: number;
+  onRefreshDefaultModels: () => void;
+  codexPathDraft: string;
+  codexArgsDraft: string;
+  codexDirty: boolean;
+  isSavingSettings: boolean;
+  doctorState: {
+    status: "idle" | "running" | "done";
+    result: CodexDoctorResult | null;
+  };
+  codexUpdateState: {
+    status: "idle" | "running" | "done";
+    result: CodexUpdateResult | null;
+  };
+  globalAgentsMeta: string;
+  globalAgentsError: string | null;
+  globalAgentsContent: string;
+  globalAgentsLoading: boolean;
+  globalAgentsRefreshDisabled: boolean;
+  globalAgentsSaveDisabled: boolean;
+  globalAgentsSaveLabel: string;
+  globalConfigMeta: string;
+  globalConfigError: string | null;
+  globalConfigContent: string;
+  globalConfigLoading: boolean;
+  globalConfigRefreshDisabled: boolean;
+  globalConfigSaveDisabled: boolean;
+  globalConfigSaveLabel: string;
+  projects: WorkspaceInfo[];
+  codexBinOverrideDrafts: Record<string, string>;
+  codexHomeOverrideDrafts: Record<string, string>;
+  codexArgsOverrideDrafts: Record<string, string>;
+  onSetCodexPathDraft: Dispatch<SetStateAction<string>>;
+  onSetCodexArgsDraft: Dispatch<SetStateAction<string>>;
+  onSetGlobalAgentsContent: (value: string) => void;
+  onSetGlobalConfigContent: (value: string) => void;
+  onSetCodexBinOverrideDrafts: Dispatch<SetStateAction<Record<string, string>>>;
+  onSetCodexHomeOverrideDrafts: Dispatch<SetStateAction<Record<string, string>>>;
+  onSetCodexArgsOverrideDrafts: Dispatch<SetStateAction<Record<string, string>>>;
+  onBrowseCodex: () => Promise<void>;
+  onSaveCodexSettings: () => Promise<void>;
+  onRunDoctor: () => Promise<void>;
+  onRunCodexUpdate: () => Promise<void>;
+  onRefreshGlobalAgents: () => void;
+  onSaveGlobalAgents: () => void;
+  onRefreshGlobalConfig: () => void;
+  onSaveGlobalConfig: () => void;
+  onUpdateWorkspaceCodexBin: (id: string, codexBin: string | null) => Promise<void>;
+  onUpdateWorkspaceSettings: (
+    id: string,
+    settings: Partial<WorkspaceSettings>,
+  ) => Promise<void>;
+};
+
+export const useSettingsCodexSection = ({
+  appSettings,
+  projects,
+  onUpdateAppSettings,
+  onRunDoctor,
+  onRunCodexUpdate,
+  onUpdateWorkspaceCodexBin,
+  onUpdateWorkspaceSettings,
+}: UseSettingsCodexSectionArgs): SettingsCodexSectionProps => {
+  const [codexPathDraft, setCodexPathDraft] = useState(appSettings.codexBin ?? "");
+  const [codexArgsDraft, setCodexArgsDraft] = useState(appSettings.codexArgs ?? "");
+  const [codexBinOverrideDrafts, setCodexBinOverrideDrafts] = useState<
+    Record<string, string>
+  >({});
+  const [codexHomeOverrideDrafts, setCodexHomeOverrideDrafts] = useState<
+    Record<string, string>
+  >({});
+  const [codexArgsOverrideDrafts, setCodexArgsOverrideDrafts] = useState<
+    Record<string, string>
+  >({});
+  const [isSavingSettings, setIsSavingSettings] = useState(false);
+  const [doctorState, setDoctorState] = useState<{
+    status: "idle" | "running" | "done";
+    result: CodexDoctorResult | null;
+  }>({ status: "idle", result: null });
+  const [codexUpdateState, setCodexUpdateState] = useState<{
+    status: "idle" | "running" | "done";
+    result: CodexUpdateResult | null;
+  }>({ status: "idle", result: null });
+
+  const {
+    models: defaultModels,
+    isLoading: defaultModelsLoading,
+    error: defaultModelsError,
+    connectedWorkspaceCount: defaultModelsConnectedWorkspaceCount,
+    refresh: refreshDefaultModels,
+  } = useSettingsDefaultModels(projects);
+
+  const {
+    content: globalAgentsContent,
+    exists: globalAgentsExists,
+    truncated: globalAgentsTruncated,
+    isLoading: globalAgentsLoading,
+    isSaving: globalAgentsSaving,
+    error: globalAgentsError,
+    isDirty: globalAgentsDirty,
+    setContent: setGlobalAgentsContent,
+    refresh: refreshGlobalAgents,
+    save: saveGlobalAgents,
+  } = useGlobalAgentsMd();
+
+  const {
+    content: globalConfigContent,
+    exists: globalConfigExists,
+    truncated: globalConfigTruncated,
+    isLoading: globalConfigLoading,
+    isSaving: globalConfigSaving,
+    error: globalConfigError,
+    isDirty: globalConfigDirty,
+    setContent: setGlobalConfigContent,
+    refresh: refreshGlobalConfig,
+    save: saveGlobalConfig,
+  } = useGlobalCodexConfigToml();
+
+  const globalAgentsEditorMeta = buildEditorContentMeta({
+    isLoading: globalAgentsLoading,
+    isSaving: globalAgentsSaving,
+    exists: globalAgentsExists,
+    truncated: globalAgentsTruncated,
+    isDirty: globalAgentsDirty,
+  });
+
+  const globalConfigEditorMeta = buildEditorContentMeta({
+    isLoading: globalConfigLoading,
+    isSaving: globalConfigSaving,
+    exists: globalConfigExists,
+    truncated: globalConfigTruncated,
+    isDirty: globalConfigDirty,
+  });
+
+  useEffect(() => {
+    setCodexPathDraft(appSettings.codexBin ?? "");
+  }, [appSettings.codexBin]);
+
+  useEffect(() => {
+    setCodexArgsDraft(appSettings.codexArgs ?? "");
+  }, [appSettings.codexArgs]);
+
+  useEffect(() => {
+    setCodexBinOverrideDrafts((prev) =>
+      buildWorkspaceOverrideDrafts(projects, prev, (workspace) => workspace.codex_bin ?? null),
+    );
+    setCodexHomeOverrideDrafts((prev) =>
+      buildWorkspaceOverrideDrafts(
+        projects,
+        prev,
+        (workspace) => workspace.settings.codexHome ?? null,
+      ),
+    );
+    setCodexArgsOverrideDrafts((prev) =>
+      buildWorkspaceOverrideDrafts(
+        projects,
+        prev,
+        (workspace) => workspace.settings.codexArgs ?? null,
+      ),
+    );
+  }, [projects]);
+
+  const nextCodexBin = codexPathDraft.trim() ? codexPathDraft.trim() : null;
+  const nextCodexArgs = codexArgsDraft.trim() ? codexArgsDraft.trim() : null;
+  const codexDirty =
+    nextCodexBin !== (appSettings.codexBin ?? null) ||
+    nextCodexArgs !== (appSettings.codexArgs ?? null);
+
+  const handleBrowseCodex = async () => {
+    const selection = await open({ multiple: false, directory: false });
+    if (!selection || Array.isArray(selection)) {
+      return;
+    }
+    setCodexPathDraft(selection);
+  };
+
+  const handleSaveCodexSettings = async () => {
+    setIsSavingSettings(true);
+    try {
+      await onUpdateAppSettings({
+        ...appSettings,
+        codexBin: nextCodexBin,
+        codexArgs: nextCodexArgs,
+      });
+    } finally {
+      setIsSavingSettings(false);
+    }
+  };
+
+  const handleRunDoctor = async () => {
+    setDoctorState({ status: "running", result: null });
+    try {
+      const result = await onRunDoctor(nextCodexBin, nextCodexArgs);
+      setDoctorState({ status: "done", result });
+    } catch (error) {
+      setDoctorState({
+        status: "done",
+        result: {
+          ok: false,
+          codexBin: nextCodexBin,
+          version: null,
+          appServerOk: false,
+          details: error instanceof Error ? error.message : String(error),
+          path: null,
+          nodeOk: false,
+          nodeVersion: null,
+          nodeDetails: null,
+        },
+      });
+    }
+  };
+
+  const handleRunCodexUpdate = async () => {
+    setCodexUpdateState({ status: "running", result: null });
+    try {
+      if (!onRunCodexUpdate) {
+        setCodexUpdateState({
+          status: "done",
+          result: {
+            ok: false,
+            method: "unknown",
+            package: null,
+            beforeVersion: null,
+            afterVersion: null,
+            upgraded: false,
+            output: null,
+            details: "Codex updates are not available in this build.",
+          },
+        });
+        return;
+      }
+
+      const result = await onRunCodexUpdate(nextCodexBin, nextCodexArgs);
+      setCodexUpdateState({ status: "done", result });
+    } catch (error) {
+      setCodexUpdateState({
+        status: "done",
+        result: {
+          ok: false,
+          method: "unknown",
+          package: null,
+          beforeVersion: null,
+          afterVersion: null,
+          upgraded: false,
+          output: null,
+          details: error instanceof Error ? error.message : String(error),
+        },
+      });
+    }
+  };
+
+  return {
+    appSettings,
+    onUpdateAppSettings,
+    defaultModels,
+    defaultModelsLoading,
+    defaultModelsError,
+    defaultModelsConnectedWorkspaceCount,
+    onRefreshDefaultModels: () => {
+      void refreshDefaultModels();
+    },
+    codexPathDraft,
+    codexArgsDraft,
+    codexDirty,
+    isSavingSettings,
+    doctorState,
+    codexUpdateState,
+    globalAgentsMeta: globalAgentsEditorMeta.meta,
+    globalAgentsError,
+    globalAgentsContent,
+    globalAgentsLoading,
+    globalAgentsRefreshDisabled: globalAgentsEditorMeta.refreshDisabled,
+    globalAgentsSaveDisabled: globalAgentsEditorMeta.saveDisabled,
+    globalAgentsSaveLabel: globalAgentsEditorMeta.saveLabel,
+    globalConfigMeta: globalConfigEditorMeta.meta,
+    globalConfigError,
+    globalConfigContent,
+    globalConfigLoading,
+    globalConfigRefreshDisabled: globalConfigEditorMeta.refreshDisabled,
+    globalConfigSaveDisabled: globalConfigEditorMeta.saveDisabled,
+    globalConfigSaveLabel: globalConfigEditorMeta.saveLabel,
+    projects,
+    codexBinOverrideDrafts,
+    codexHomeOverrideDrafts,
+    codexArgsOverrideDrafts,
+    onSetCodexPathDraft: setCodexPathDraft,
+    onSetCodexArgsDraft: setCodexArgsDraft,
+    onSetGlobalAgentsContent: setGlobalAgentsContent,
+    onSetGlobalConfigContent: setGlobalConfigContent,
+    onSetCodexBinOverrideDrafts: setCodexBinOverrideDrafts,
+    onSetCodexHomeOverrideDrafts: setCodexHomeOverrideDrafts,
+    onSetCodexArgsOverrideDrafts: setCodexArgsOverrideDrafts,
+    onBrowseCodex: handleBrowseCodex,
+    onSaveCodexSettings: handleSaveCodexSettings,
+    onRunDoctor: handleRunDoctor,
+    onRunCodexUpdate: handleRunCodexUpdate,
+    onRefreshGlobalAgents: () => {
+      void refreshGlobalAgents();
+    },
+    onSaveGlobalAgents: () => {
+      void saveGlobalAgents();
+    },
+    onRefreshGlobalConfig: () => {
+      void refreshGlobalConfig();
+    },
+    onSaveGlobalConfig: () => {
+      void saveGlobalConfig();
+    },
+    onUpdateWorkspaceCodexBin,
+    onUpdateWorkspaceSettings,
+  };
+};
