+++ src/features/settings/hooks/useSettingsDisplaySection.ts
@@ -1,0 +1,173 @@
+import { useEffect, useState } from "react";
+import type { Dispatch, SetStateAction } from "react";
+import type { AppSettings } from "@/types";
+import { clampUiScale } from "@utils/uiScale";
+import {
+  DEFAULT_CODE_FONT_FAMILY,
+  DEFAULT_UI_FONT_FAMILY,
+  clampCodeFontSize,
+  normalizeFontFamily,
+} from "@utils/fonts";
+
+type UseSettingsDisplaySectionArgs = {
+  appSettings: AppSettings;
+  reduceTransparency: boolean;
+  onToggleTransparency: (value: boolean) => void;
+  onUpdateAppSettings: (next: AppSettings) => Promise<void>;
+  scaleShortcutTitle: string;
+  scaleShortcutText: string;
+  onTestNotificationSound: () => void;
+  onTestSystemNotification: () => void;
+};
+
+export type SettingsDisplaySectionProps = {
+  appSettings: AppSettings;
+  reduceTransparency: boolean;
+  scaleShortcutTitle: string;
+  scaleShortcutText: string;
+  scaleDraft: string;
+  uiFontDraft: string;
+  codeFontDraft: string;
+  codeFontSizeDraft: number;
+  onUpdateAppSettings: (next: AppSettings) => Promise<void>;
+  onToggleTransparency: (value: boolean) => void;
+  onSetScaleDraft: Dispatch<SetStateAction<string>>;
+  onCommitScale: () => Promise<void>;
+  onResetScale: () => Promise<void>;
+  onSetUiFontDraft: Dispatch<SetStateAction<string>>;
+  onCommitUiFont: () => Promise<void>;
+  onSetCodeFontDraft: Dispatch<SetStateAction<string>>;
+  onCommitCodeFont: () => Promise<void>;
+  onSetCodeFontSizeDraft: Dispatch<SetStateAction<number>>;
+  onCommitCodeFontSize: (nextSize: number) => Promise<void>;
+  onTestNotificationSound: () => void;
+  onTestSystemNotification: () => void;
+};
+
+export const useSettingsDisplaySection = ({
+  appSettings,
+  reduceTransparency,
+  onToggleTransparency,
+  onUpdateAppSettings,
+  scaleShortcutTitle,
+  scaleShortcutText,
+  onTestNotificationSound,
+  onTestSystemNotification,
+}: UseSettingsDisplaySectionArgs): SettingsDisplaySectionProps => {
+  const [scaleDraft, setScaleDraft] = useState(
+    `${Math.round(clampUiScale(appSettings.uiScale) * 100)}%`,
+  );
+  const [uiFontDraft, setUiFontDraft] = useState(appSettings.uiFontFamily);
+  const [codeFontDraft, setCodeFontDraft] = useState(appSettings.codeFontFamily);
+  const [codeFontSizeDraft, setCodeFontSizeDraft] = useState(appSettings.codeFontSize);
+
+  useEffect(() => {
+    setScaleDraft(`${Math.round(clampUiScale(appSettings.uiScale) * 100)}%`);
+  }, [appSettings.uiScale]);
+
+  useEffect(() => {
+    setUiFontDraft(appSettings.uiFontFamily);
+  }, [appSettings.uiFontFamily]);
+
+  useEffect(() => {
+    setCodeFontDraft(appSettings.codeFontFamily);
+  }, [appSettings.codeFontFamily]);
+
+  useEffect(() => {
+    setCodeFontSizeDraft(appSettings.codeFontSize);
+  }, [appSettings.codeFontSize]);
+
+  const trimmedScale = scaleDraft.trim();
+  const parsedPercent = trimmedScale
+    ? Number(trimmedScale.replace("%", ""))
+    : Number.NaN;
+  const parsedScale = Number.isFinite(parsedPercent) ? parsedPercent / 100 : null;
+
+  const handleCommitScale = async () => {
+    if (parsedScale === null) {
+      setScaleDraft(`${Math.round(clampUiScale(appSettings.uiScale) * 100)}%`);
+      return;
+    }
+    const nextScale = clampUiScale(parsedScale);
+    setScaleDraft(`${Math.round(nextScale * 100)}%`);
+    if (nextScale === appSettings.uiScale) {
+      return;
+    }
+    await onUpdateAppSettings({
+      ...appSettings,
+      uiScale: nextScale,
+    });
+  };
+
+  const handleResetScale = async () => {
+    if (appSettings.uiScale === 1) {
+      setScaleDraft("100%");
+      return;
+    }
+    setScaleDraft("100%");
+    await onUpdateAppSettings({
+      ...appSettings,
+      uiScale: 1,
+    });
+  };
+
+  const handleCommitUiFont = async () => {
+    const nextFont = normalizeFontFamily(uiFontDraft, DEFAULT_UI_FONT_FAMILY);
+    setUiFontDraft(nextFont);
+    if (nextFont === appSettings.uiFontFamily) {
+      return;
+    }
+    await onUpdateAppSettings({
+      ...appSettings,
+      uiFontFamily: nextFont,
+    });
+  };
+
+  const handleCommitCodeFont = async () => {
+    const nextFont = normalizeFontFamily(codeFontDraft, DEFAULT_CODE_FONT_FAMILY);
+    setCodeFontDraft(nextFont);
+    if (nextFont === appSettings.codeFontFamily) {
+      return;
+    }
+    await onUpdateAppSettings({
+      ...appSettings,
+      codeFontFamily: nextFont,
+    });
+  };
+
+  const handleCommitCodeFontSize = async (nextSize: number) => {
+    const clampedSize = clampCodeFontSize(nextSize);
+    setCodeFontSizeDraft(clampedSize);
+    if (clampedSize === appSettings.codeFontSize) {
+      return;
+    }
+    await onUpdateAppSettings({
+      ...appSettings,
+      codeFontSize: clampedSize,
+    });
+  };
+
+  return {
+    appSettings,
+    reduceTransparency,
+    scaleShortcutTitle,
+    scaleShortcutText,
+    scaleDraft,
+    uiFontDraft,
+    codeFontDraft,
+    codeFontSizeDraft,
+    onUpdateAppSettings,
+    onToggleTransparency,
+    onSetScaleDraft: setScaleDraft,
+    onCommitScale: handleCommitScale,
+    onResetScale: handleResetScale,
+    onSetUiFontDraft: setUiFontDraft,
+    onCommitUiFont: handleCommitUiFont,
+    onSetCodeFontDraft: setCodeFontDraft,
+    onCommitCodeFont: handleCommitCodeFont,
+    onSetCodeFontSizeDraft: setCodeFontSizeDraft,
+    onCommitCodeFontSize: handleCommitCodeFontSize,
+    onTestNotificationSound,
+    onTestSystemNotification,
+  };
+};
