+++ src/features/settings/hooks/useSettingsEnvironmentsSection.ts
@@ -1,0 +1,141 @@
+import { useEffect, useMemo, useState } from "react";
+import type { Dispatch, SetStateAction } from "react";
+import type { WorkspaceInfo } from "@/types";
+import { normalizeWorktreeSetupScript } from "@settings/components/settingsViewHelpers";
+
+type UseSettingsEnvironmentsSectionArgs = {
+  mainWorkspaces: WorkspaceInfo[];
+  onUpdateWorkspaceSettings: (
+    id: string,
+    settings: Partial<WorkspaceInfo["settings"]>,
+  ) => Promise<void>;
+};
+
+export type SettingsEnvironmentsSectionProps = {
+  mainWorkspaces: WorkspaceInfo[];
+  environmentWorkspace: WorkspaceInfo | null;
+  environmentSaving: boolean;
+  environmentError: string | null;
+  environmentDraftScript: string;
+  environmentSavedScript: string | null;
+  environmentDirty: boolean;
+  onSetEnvironmentWorkspaceId: Dispatch<SetStateAction<string | null>>;
+  onSetEnvironmentDraftScript: Dispatch<SetStateAction<string>>;
+  onSaveEnvironmentSetup: () => Promise<void>;
+};
+
+export const useSettingsEnvironmentsSection = ({
+  mainWorkspaces,
+  onUpdateWorkspaceSettings,
+}: UseSettingsEnvironmentsSectionArgs): SettingsEnvironmentsSectionProps => {
+  const [environmentWorkspaceId, setEnvironmentWorkspaceId] = useState<string | null>(
+    null,
+  );
+  const [environmentDraftScript, setEnvironmentDraftScript] = useState("");
+  const [environmentSavedScript, setEnvironmentSavedScript] = useState<string | null>(
+    null,
+  );
+  const [environmentLoadedWorkspaceId, setEnvironmentLoadedWorkspaceId] = useState<
+    string | null
+  >(null);
+  const [environmentError, setEnvironmentError] = useState<string | null>(null);
+  const [environmentSaving, setEnvironmentSaving] = useState(false);
+
+  const environmentWorkspace = useMemo(() => {
+    if (mainWorkspaces.length === 0) {
+      return null;
+    }
+    if (environmentWorkspaceId) {
+      const found = mainWorkspaces.find((workspace) => workspace.id === environmentWorkspaceId);
+      if (found) {
+        return found;
+      }
+    }
+    return mainWorkspaces[0] ?? null;
+  }, [environmentWorkspaceId, mainWorkspaces]);
+
+  const environmentSavedScriptFromWorkspace = useMemo(() => {
+    return normalizeWorktreeSetupScript(environmentWorkspace?.settings.worktreeSetupScript);
+  }, [environmentWorkspace?.settings.worktreeSetupScript]);
+
+  const environmentDraftNormalized = useMemo(() => {
+    return normalizeWorktreeSetupScript(environmentDraftScript);
+  }, [environmentDraftScript]);
+
+  const environmentDirty = environmentDraftNormalized !== environmentSavedScript;
+
+  useEffect(() => {
+    if (!environmentWorkspace) {
+      setEnvironmentWorkspaceId(null);
+      setEnvironmentLoadedWorkspaceId(null);
+      setEnvironmentSavedScript(null);
+      setEnvironmentDraftScript("");
+      setEnvironmentError(null);
+      setEnvironmentSaving(false);
+      return;
+    }
+
+    if (environmentWorkspaceId !== environmentWorkspace.id) {
+      setEnvironmentWorkspaceId(environmentWorkspace.id);
+    }
+  }, [environmentWorkspace, environmentWorkspaceId]);
+
+  useEffect(() => {
+    if (!environmentWorkspace) {
+      return;
+    }
+
+    if (environmentLoadedWorkspaceId !== environmentWorkspace.id) {
+      setEnvironmentLoadedWorkspaceId(environmentWorkspace.id);
+      setEnvironmentSavedScript(environmentSavedScriptFromWorkspace);
+      setEnvironmentDraftScript(environmentSavedScriptFromWorkspace ?? "");
+      setEnvironmentError(null);
+      return;
+    }
+
+    if (!environmentDirty && environmentSavedScript !== environmentSavedScriptFromWorkspace) {
+      setEnvironmentSavedScript(environmentSavedScriptFromWorkspace);
+      setEnvironmentDraftScript(environmentSavedScriptFromWorkspace ?? "");
+      setEnvironmentError(null);
+    }
+  }, [
+    environmentDirty,
+    environmentLoadedWorkspaceId,
+    environmentSavedScript,
+    environmentSavedScriptFromWorkspace,
+    environmentWorkspace,
+  ]);
+
+  const handleSaveEnvironmentSetup = async () => {
+    if (!environmentWorkspace || environmentSaving) {
+      return;
+    }
+    const nextScript = environmentDraftNormalized;
+    setEnvironmentSaving(true);
+    setEnvironmentError(null);
+    try {
+      await onUpdateWorkspaceSettings(environmentWorkspace.id, {
+        worktreeSetupScript: nextScript,
+      });
+      setEnvironmentSavedScript(nextScript);
+      setEnvironmentDraftScript(nextScript ?? "");
+    } catch (error) {
+      setEnvironmentError(error instanceof Error ? error.message : String(error));
+    } finally {
+      setEnvironmentSaving(false);
+    }
+  };
+
+  return {
+    mainWorkspaces,
+    environmentWorkspace,
+    environmentSaving,
+    environmentError,
+    environmentDraftScript,
+    environmentSavedScript,
+    environmentDirty,
+    onSetEnvironmentWorkspaceId: setEnvironmentWorkspaceId,
+    onSetEnvironmentDraftScript: setEnvironmentDraftScript,
+    onSaveEnvironmentSetup: handleSaveEnvironmentSetup,
+  };
+};
