+++ src/features/settings/hooks/useSettingsProjectsSection.ts
@@ -1,0 +1,194 @@
+import { useEffect, useMemo, useState } from "react";
+import type { Dispatch, SetStateAction } from "react";
+import { ask, open } from "@tauri-apps/plugin-dialog";
+import type { AppSettings, WorkspaceGroup, WorkspaceInfo } from "@/types";
+import type { GroupedWorkspaces } from "./settingsSectionTypes";
+
+type UseSettingsProjectsSectionArgs = {
+  appSettings: AppSettings;
+  workspaceGroups: WorkspaceGroup[];
+  groupedWorkspaces: GroupedWorkspaces;
+  ungroupedLabel: string;
+  projects: WorkspaceInfo[];
+  onUpdateAppSettings: (next: AppSettings) => Promise<void>;
+  onMoveWorkspace: (id: string, direction: "up" | "down") => void;
+  onDeleteWorkspace: (id: string) => void;
+  onCreateWorkspaceGroup: (name: string) => Promise<WorkspaceGroup | null>;
+  onRenameWorkspaceGroup: (id: string, name: string) => Promise<boolean | null>;
+  onMoveWorkspaceGroup: (id: string, direction: "up" | "down") => Promise<boolean | null>;
+  onDeleteWorkspaceGroup: (id: string) => Promise<boolean | null>;
+  onAssignWorkspaceGroup: (
+    workspaceId: string,
+    groupId: string | null,
+  ) => Promise<boolean | null>;
+};
+
+export type SettingsProjectsSectionProps = {
+  workspaceGroups: WorkspaceGroup[];
+  groupedWorkspaces: GroupedWorkspaces;
+  ungroupedLabel: string;
+  groupDrafts: Record<string, string>;
+  newGroupName: string;
+  groupError: string | null;
+  projects: WorkspaceInfo[];
+  canCreateGroup: boolean;
+  onSetNewGroupName: Dispatch<SetStateAction<string>>;
+  onSetGroupDrafts: Dispatch<SetStateAction<Record<string, string>>>;
+  onCreateGroup: () => Promise<void>;
+  onRenameGroup: (group: WorkspaceGroup) => Promise<void>;
+  onMoveWorkspaceGroup: (id: string, direction: "up" | "down") => Promise<boolean | null>;
+  onDeleteGroup: (group: WorkspaceGroup) => Promise<void>;
+  onChooseGroupCopiesFolder: (group: WorkspaceGroup) => Promise<void>;
+  onClearGroupCopiesFolder: (group: WorkspaceGroup) => Promise<void>;
+  onAssignWorkspaceGroup: (
+    workspaceId: string,
+    groupId: string | null,
+  ) => Promise<boolean | null>;
+  onMoveWorkspace: (id: string, direction: "up" | "down") => void;
+  onDeleteWorkspace: (id: string) => void;
+};
+
+export const useSettingsProjectsSection = ({
+  appSettings,
+  workspaceGroups,
+  groupedWorkspaces,
+  ungroupedLabel,
+  projects,
+  onUpdateAppSettings,
+  onMoveWorkspace,
+  onDeleteWorkspace,
+  onCreateWorkspaceGroup,
+  onRenameWorkspaceGroup,
+  onMoveWorkspaceGroup,
+  onDeleteWorkspaceGroup,
+  onAssignWorkspaceGroup,
+}: UseSettingsProjectsSectionArgs): SettingsProjectsSectionProps => {
+  const [groupDrafts, setGroupDrafts] = useState<Record<string, string>>({});
+  const [newGroupName, setNewGroupName] = useState("");
+  const [groupError, setGroupError] = useState<string | null>(null);
+
+  useEffect(() => {
+    setGroupDrafts((prev) => {
+      const next: Record<string, string> = {};
+      workspaceGroups.forEach((group) => {
+        next[group.id] = prev[group.id] ?? group.name;
+      });
+      return next;
+    });
+  }, [workspaceGroups]);
+
+  const trimmedGroupName = useMemo(() => newGroupName.trim(), [newGroupName]);
+  const canCreateGroup = Boolean(trimmedGroupName);
+
+  const handleCreateGroup = async () => {
+    setGroupError(null);
+    try {
+      const created = await onCreateWorkspaceGroup(newGroupName);
+      if (created) {
+        setNewGroupName("");
+      }
+    } catch (error) {
+      setGroupError(error instanceof Error ? error.message : String(error));
+    }
+  };
+
+  const handleRenameGroup = async (group: WorkspaceGroup) => {
+    const draft = groupDrafts[group.id] ?? "";
+    const trimmed = draft.trim();
+    if (!trimmed || trimmed === group.name) {
+      setGroupDrafts((prev) => ({
+        ...prev,
+        [group.id]: group.name,
+      }));
+      return;
+    }
+    setGroupError(null);
+    try {
+      await onRenameWorkspaceGroup(group.id, trimmed);
+    } catch (error) {
+      setGroupError(error instanceof Error ? error.message : String(error));
+      setGroupDrafts((prev) => ({
+        ...prev,
+        [group.id]: group.name,
+      }));
+    }
+  };
+
+  const updateGroupCopiesFolder = async (
+    groupId: string,
+    copiesFolder: string | null,
+  ) => {
+    setGroupError(null);
+    try {
+      await onUpdateAppSettings({
+        ...appSettings,
+        workspaceGroups: appSettings.workspaceGroups.map((entry) =>
+          entry.id === groupId ? { ...entry, copiesFolder } : entry,
+        ),
+      });
+    } catch (error) {
+      setGroupError(error instanceof Error ? error.message : String(error));
+    }
+  };
+
+  const handleChooseGroupCopiesFolder = async (group: WorkspaceGroup) => {
+    const selection = await open({ multiple: false, directory: true });
+    if (!selection || Array.isArray(selection)) {
+      return;
+    }
+    await updateGroupCopiesFolder(group.id, selection);
+  };
+
+  const handleClearGroupCopiesFolder = async (group: WorkspaceGroup) => {
+    if (!group.copiesFolder) {
+      return;
+    }
+    await updateGroupCopiesFolder(group.id, null);
+  };
+
+  const handleDeleteGroup = async (group: WorkspaceGroup) => {
+    const groupProjects =
+      groupedWorkspaces.find((entry) => entry.id === group.id)?.workspaces ?? [];
+    const detail =
+      groupProjects.length > 0
+        ? `\n\nProjects in this group will move to "${ungroupedLabel}".`
+        : "";
+    const confirmed = await ask(`Delete "${group.name}"?${detail}`, {
+      title: "Delete Group",
+      kind: "warning",
+      okLabel: "Delete",
+      cancelLabel: "Cancel",
+    });
+    if (!confirmed) {
+      return;
+    }
+    setGroupError(null);
+    try {
+      await onDeleteWorkspaceGroup(group.id);
+    } catch (error) {
+      setGroupError(error instanceof Error ? error.message : String(error));
+    }
+  };
+
+  return {
+    workspaceGroups,
+    groupedWorkspaces,
+    ungroupedLabel,
+    groupDrafts,
+    newGroupName,
+    groupError,
+    projects,
+    canCreateGroup,
+    onSetNewGroupName: setNewGroupName,
+    onSetGroupDrafts: setGroupDrafts,
+    onCreateGroup: handleCreateGroup,
+    onRenameGroup: handleRenameGroup,
+    onMoveWorkspaceGroup,
+    onDeleteGroup: handleDeleteGroup,
+    onChooseGroupCopiesFolder: handleChooseGroupCopiesFolder,
+    onClearGroupCopiesFolder: handleClearGroupCopiesFolder,
+    onAssignWorkspaceGroup,
+    onMoveWorkspace,
+    onDeleteWorkspace,
+  };
+};
